RomainSARRAZIN_HarenaRAKOTOMANGA_PostsReseaux/groupe-social-media-creator/requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
pydantic>=2.5.0
requests>=2.31.0
Pillow>=10.0.0
